Change examples' submodules
Tracking CASE-Examples-QC as a submodule causes a circular Git submodule
loop between `case-prov`, CASE-Corpora, and CASE-Examples-QC.
CASE-Examples-QC was being tracked to keep a consistent tracking status
between CASE-Examples and the CASE website.

This patch moves to tracking CASE-Examples and the CASE website
directly, as the synchronization mechanism is now believed to be able to
be handled by a `check-supply-chain` Makefile target, like is currently
used with CASE-Corpora and other repositories.

This patch only changes the submodule tracking links, leaving the
tracked versions alone.  Follow-on patches will update the submodules'
states and add the `check-supply-chain` review workflow.

No effects were observed on Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>